Try to fix MacOS first tab on Chrome launch
As it seems (can't reproduce - no Mac here) users with MacOS have their
most visited still visible on the first Chrome launch.

My best guess is, that the extension fires too early (at this time the
most visited tiles might not be there) and then, when they are loaded,
the extension already tried to remove them.

I added an optional callback to the `modEl` (modify Element) function,
which and when the catch clause is reached I retry it for 10 times and
wait 200 ms between each retry. This is a really dirty workaround, but
hopefully this fixes the MacOS issue.